1.12.2-1.0.1
ported the entire mod to 1.12.2.
Removed the toast. just didn't want to deal with it.
Removed code that the original author didn't even address.
Removed that was not really needed because of the way recipes are handled.
Made recipes a little more complicated, but not by much and it's acceptable really. REMEMBER to use JEI.